Remove system() calls from callbacks.c. (NF 021021)
git-svn-id: svn://projects.linuxtogo.org/svn/gpe/trunk@1841 d4735bf7-8e1f-0410-bce4-ef584eb11c25

+++ base/gpe-announce/callbacks.c
@@ -47,16 +47,18 @@ char filename[256], command[256];
 FILE *f;
 	
 	sprintf(filename, "/var/spool/at/%d.1234", (int)when);
-  
+	f=NULL;
 	f=fopen(filename, "w");
-	
-	fprintf(f, "#!/bin/sh\n");
-	fprintf(f, "export DISPLAY=:0.0\n");
-	fprintf(f, "/usr/bin/gpe-announce '%s'\n", buf);
-	fprintf(f, "/bin/rm $0\n");
+	if (f != NULL) {
+		fprintf(f, "#!/bin/sh\n");
+		fprintf(f, "export DISPLAY=:0.0\n");
+		fprintf(f, "/usr/bin/gpe-announce '%s'\n", buf);
+		fprintf(f, "/bin/rm $0\n");
+		fclose(f);
+	} else
+		return;
   
-	fclose(f);
-  
+#if 0
 	sprintf(command, "chmod 755 %s", filename);
   		  
 	system(command);
@@ -64,7 +66,15 @@ FILE *f;
 	sprintf(command, "echo >/var/spool/at/trigger");
   		  
 	system(command);
-  
+#else
+	chmod (filename, S_IRUSR | S_IWUSR | S_IXUSR | S_IRGRP | S_IXGRP | S_IROTH | S_IXOTH);
+	f=NULL;
+	f=fopen("/var/spool/at/trigger","w");
+	if (f != NULL) {
+		fprintf(f,"\n");
+		fclose(f);
+	}
+#endif
 }
 	
 void play_melody(guint Tone1Pitch, guint Tone1Duration,
